+++ pysatSeasons/plot.py
@@ -1,5 +1,3 @@
-from __future__ import print_function
-from __future__ import absolute_import
 
 import matplotlib as mpl
 import matplotlib.pyplot as plt